Move all decomposition related ops into separate file
