+++ __init__.py
@@ -4,3 +4,4 @@ from ops import *
 from initializers import *
 from regularizers import *
 from shapes import *
+from decompositions import *
