+++ ops_test.py
@@ -3,7 +3,6 @@ import tensorflow as tf
 
 import tensor_train
 import ops
-import shapes
 import initializers
 
 
@@ -36,27 +35,6 @@ class TTTensorTest(tf.test.TestCase):
         tf_tens = tensor_train.TensorTrain(tt_cores)
         actual = ops.full(tf_tens)
         self.assertAllClose(desired, actual.eval())
-
-  def testTTTensor(self):
-    shape = (2, 1, 4, 3)
-    np.random.seed(1)
-    tens = np.random.rand(*shape).astype(np.float32)
-    tf_tens = tf.constant(tens)
-    tt_tens = ops.to_tt_tensor(tf_tens, max_tt_rank=3)
-    with self.test_session():
-      self.assertAllClose(tens, ops.full(tt_tens).eval())
-      dynamic_tt_ranks = shapes.tt_ranks(tt_tens).eval()
-      static_tt_ranks = tt_tens.get_tt_ranks().as_list()
-      self.assertAllEqual(dynamic_tt_ranks, static_tt_ranks)
-
-      # Try to decompose the same tensor with unknown shape.
-      tf_tens_pl = tf.placeholder(tf.float32, (None, None, 4, None))
-      tt_tens = ops.to_tt_tensor(tf_tens_pl, max_tt_rank=3)
-      tt_val = ops.full(tt_tens).eval({tf_tens_pl: tens})
-      self.assertAllClose(tens, tt_val)
-      dynamic_tt_ranks = shapes.tt_ranks(tt_tens).eval({tf_tens_pl: tens})
-      self.assertAllEqual(dynamic_tt_ranks, static_tt_ranks)
-
 
 
 class TTMatrixTest(tf.test.TestCase):
@@ -95,29 +73,6 @@ class TTMatrixTest(tf.test.TestCase):
         tf_mat = tensor_train.TensorTrain(tt_cores)
         actual = ops.full(tf_mat)
         self.assertAllClose(desired, actual.eval())
-
-  def testTTVector(self):
-    vec_shape = (2, 1, 4, 3)
-    np.random.seed(1)
-    rows = np.prod(vec_shape)
-    vec = np.random.rand(rows, 1).astype(np.float32)
-    with self.test_session():
-      tf_vec = tf.constant(vec)
-      tt_vec = ops.to_tt_matrix(tf_vec, (vec_shape, None))
-      self.assertAllClose(vec, ops.full(tt_vec).eval())
-
-  def testTTMatrix(self):
-    # Convert a np.prod(out_shape) x np.prod(in_shape) matrix into TT-matrix
-    # and back.
-    inp_shape = (2, 5, 2, 3)
-    out_shape = (3, 3, 2, 3)
-    np.random.seed(1)
-    mat = np.random.rand(np.prod(out_shape), np.prod(inp_shape)).astype(np.float32)
-    with self.test_session():
-      tf_mat = tf.constant(mat)
-      tt_mat = ops.to_tt_matrix(tf_mat, (out_shape, inp_shape), max_tt_rank=90)
-      # TODO: why so bad accuracy?
-      self.assertAllClose(mat, ops.full(tt_mat).eval(), atol=1e-5, rtol=1e-5)
 
   def testTTMatTimesTTMat(self):
     # Multiply a TT-matrix by another TT-matrix.
